ENH: add Normal Inverse Gaussian (nig) distribution to scipy.stats (distributions.py).
The Normal Inverse Gaussian distribution gains quite a lot of momentum in the last years as a fat tail distribution (finance, Levy processes, extreme values, etc. see http://en.wikipedia.org/wiki/Normal-inverse_Gaussian_distribution)

I am mathematician and newbee in Python. Out of curiosity I coded with scipy a method to compute risk measures with nig distribution as base. Therefore I added the nig ppf to the library of continuous distribution functions in distributions.py. This is not difficult and I copied the code used for other distributions. However, being new to python, this may be not optimal.

I crossed checked with Matlab and R and so far it matches the result obtained there.

I am still thinking about a method to compute the cdf as well as the ppf since there are no close form for them and the performance are very poor in comparison with the equivalent in R or Matlab (This is central for our applications).

Being new to python, git and all the stuff I hope that I did the things at least halfway right.

~